webservice/requirements.txt
setuptools
pip
wheel
requests==2.9.1
pyquery==1.2.17
docopt
pyyaml
